[BUG] or seems to me
As it, if i have a marking and a transition with data this doesn't work. And , this seems logical because as it, the marking should have all froms for its wanted transition. I think this is a bug but maybe it is a design's choice or other thing. 

If it's a bug, i redraw this pull request by tests, clear explinations and a reformated issue.

```yaml
Workflow.php on line 290:
Marking {#6
  -places: array:1 [▼
    "initial" => 1
  ]
}
Workflow.php on line 290:
array:1 [▼
  0 => Transition {#548 ▼
    -name: "go_search"
    -froms: array:3 [▼
      0 => "initial"
      1 => "vehicles"
      2 => "informations"
    ]
    -tos: array:1 [▼
      0 => "search"
    ]
  }
]
```